move output to bt pkg
