+++ bt/Output.go
@@ -1,4 +1,4 @@
-package output
+package bt
 
 import (
 	"encoding/binary"
